Move new series checking to new class
The series checking functionality has been moved to a new class and a
helper function for getting latest episode ID has been moved to the
Utils class.
